+++ src/orquestra/sdk/exceptions.py
@@ -3,6 +3,7 @@
 ################################################################################
 import typing as t
 
+from orquestra.sdk.schema.ir import TaskInvocationId
 from orquestra.sdk.schema.workflow_run import State
 
 
@@ -129,6 +130,16 @@ class WorkflowRunNotStarted(WorkflowRunNotFoundError):
 
 class TaskRunNotFound(NotFoundError):
     pass
+
+
+class TaskInvocationNotFoundError(NotFoundError):
+    """
+    Raised when we can't find a Task Invocation that matches the provided ID.
+    """
+
+    def __init__(self, invocation_id: TaskInvocationId):
+        super().__init__()
+        self.invocation_id = invocation_id
 
 
 # Project Errors