+++ tests/cli/dorq/test_dumpers.py
@@ -73,3 +73,38 @@ class TestArtifactDumper:
 
         # Created extension
         assert details.file_path.suffix in {".json", ".pickle"}
+
+
+class TestLogsDumper:
+    @staticmethod
+    def test_logs_dumper(tmp_path: Path):
+        # Given
+        log_values = ["my_logs", "next_log"]
+        task_invocation = "my_task_invocation"
+        logs = {task_invocation: log_values}
+        wf_run_id = "wf.1234"
+        wf_log_file = wf_run_id + ".log"
+        dir_path = tmp_path / "new_dir"
+        dumper = _dumpers.LogsDumper()
+
+        # When
+        path = dumper.dump(
+            logs=logs,
+            wf_run_id=wf_run_id,
+            dir_path=dir_path,
+        )
+
+        # Then
+        # Creates file
+        children = list(dir_path.iterdir())
+        assert len(children) == 1
+
+        # Sensible dump details
+        assert path == (dir_path / wf_log_file)
+
+        with path.open("r") as f:
+            full_logs = "".join(f.readlines())
+            for log_value in log_values:
+                # all logs should be in the file
+                assert log_value in full_logs
+            assert task_invocation in full_logs